moved more functions into person_init
